#64 마이페이지 리덕스 작업
좋아요 수, 참여중인 그룹 수 redux 작업 communitySlice건듬
